Update format before stripping cruft
I'm not using very much of this file so I've no real reason to keep it
in tact. I'm going to chop & screw the majority of it out and change the
return values to what I need shortly; in the meantime I'm just
reformatting everything to fit the style of LightFX because I'm
neurotic and take liberties with other people's code.