Refactor and merge with bfx
The blurstack algo (bfx or BlurFX) has been to this point a separate
thing. I feel like if this project grows merging the two would be a bad
idea, but for the time being that's exactly what I'm going to do. This
commit also includes further changes to the method names of LightFX.