Refactor, bugfix (stackblur, quickblur)
Added `_init` method to initialize LightFX with default image, which
saves about 3 boilerplate LOC when doing a simple blur, etc. Also
refactored `quickblur` to be canvas-mutating, while `_quickblur` is not (much
like the implementations of `stackblur` and `_stackblur`, and in line with
the convention of almost all the code in LightFX).